requirements.txt
absl-py==2.1.0
accelerate==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
bitsandbytes==0.41.1
blessed==1.20.0
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
crawlerdetect==0.1.7
cvzone==1.5.6
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
Django==5.1.2
dnspython==2.7.0
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.1.0
faiss-cpu==1.7.4
fastapi==0.115.4
fastapi-cli==0.0.5
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.11
gitignore_parser==0.1.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.19.4
idna==3.10
importlib_metadata==8.5.0
imy==0.4.0
introspection==1.9.2
ipykernel==6.29.5
ipython==8.28.0
isort==5.13.2
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jinxed==1.3.0
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
keyring==24.3.1
kiwisolver==1.4.7
langchain==0.1.4
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.5
langcodes==3.4.1
langsmith==0.0.87
language_data==1.2.0
Levenshtein==0.26.1
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.26.1
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.8
ml_dtypes==0.5.1
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.13.5
neo4j==5.14.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openai==1.73.0
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.14.1
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
parso==0.8.4
path-imports==1.1.2
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.3
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-Levenshtein==0.26.1
python-multipart==0.0.17
pytz==2024.2
pyvis==0.3.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
revel==0.9.1
rich==13.9.2
rio-ui==0.10.8
rpds-py==0.20.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentinel==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlparse==0.5.1
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.2
streamlit==1.42.1
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
sympy==1.13.3
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.5.2
timer-dict==1.0.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.13.2
torch==2.1.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.36.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.108
ultralytics-thop==2.0.14
unicall==0.1.5
uniserde==0.3.14
urllib3==2.2.3
uvicorn==0.29.0
wasabi==1.1.3
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.3.4
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.17.2
zipp==3.21.0
